evaluator: severe function call fixups
Two things were wrong.  Firstly, CALL would set the PC to the first
instruction of the called function, but the whole loop is set up so
that PC is increased after every instruction.  Thus, the first
instruction was never evaluated.

Secondly, RET has to remove the local slots and move down the return
value accordingly ('nipping', in forth speak).

Also add a sanity check on every RET that there is only one return
value so to catch compilation errors earlier.