eval: fix closure instantiation
When a closure is loaded, we have to actually duplicate the function
object and set up the foreign slots in the copy.

That this was not done in the first place is evidence of a blatant
thinko, for two reasons:

A closure being loaded several times would oops in the first place as
we would then try to set up the same foreign slot array twice.

And even if this would have worked, each load can happen with a
different stack state and absolute stack indirect pointers into the
stack are not working in this case, obviously.