object: do not free free objects on exit
Every used object needs to have a type and the only way drain_pool()
can know that an object is unused is by looking at its type pointer.

This in turn can only work if the pointer is reliably NULL for free
objects.  Make it so, or we end up calling ->type->free() on a free
object, interpreting the object pool linkage at ->data as real object
data and crash.