core: unbind function slot on compilation failure
The function slot is allocated and bound before the body is compiled
to allow self-recursive definitions.  However, if the compilation
fails, the slot is still bound to the name and a call to it crashes on
evaluation.

Unbind the slot in case of failure and further callsites will just get
'unbound name' instead.

This is an ugly fix, contributing to the uglyness of the function
compiler and it doesn't release the potentially global slot, just
removes every reference to it.