object: fix pool scanning loop
When the pool is moved to the partial list due to freeing some objects
in it, it is linked to it by means of pool->next = vm->parts.

However, the loop scans exactly that pool->next on the next iteration,
which means we end up scanning POOL_SIZE objects in a partial pool.

This, of course, crashes sooner or later, when we run into unused,
thus invalid objects and use their ->type and ->data.