object: proper pool deletion between two full pools
Currently, we remember the last pool when cycling over the full pools
to unlink a moved pool properly by making our previous pool point
properly to our next one.

However, if the previous pool was moved as well, the `last' pointer
points to a now partial pool and no longer to a node in the full pool
list.

Fix it up by first renaming those pesky variables to be more
meaningful, and then only remember the last full pool for the right
full-pool-list linkage.

Otherwise, we may write to a release pool descriptor or link a full
pool to the partial list.