compile: fix non-function scope level lookup
Checking for a local slot reference with `current == context' in the
environment lookup is wrong - we may switch non-function contexts and
still have a local slot reference.

Use the distance counter value to check for foreign references,
otherwise local slot references are detected wrongly as foreign
references with distance zero.

	(with (a 1)
	  (block
	    a))

a is a local slot reference and the code would detect it as a foreign
due to reference and owner living in two different contexts.