compile: allow for module:member names
With this in place, modules are essentially working now.  What is
still missing is implementing (variable) and (function) name binding
properly.

One artifact of the immediate compile-time loading and name resolution
for modules is the ugly semantics of changing a module slot directly
for places where usually new bindings are created (with, variable,
function).  This needs more thought.

Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>