module: do not dlclose() for now
dlclose() is a bad idea as long as there are still objects floating
around with pointers into a module's memory.

Remove it competely for now, let the kernel clean up the resources on
exit().

A cleaner long-term fix is probably to remember all modules and
deconstruct them together with the vm itself.

Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>